refactor: moves merkleTestCases.json and domainHashTestCases.json to root vectors folder
